Version 2.0.4 Build 23

+++ piwigo/ImageUploadProgressView.m
@@ -136,7 +136,7 @@
 	{
 		[self.uploadProgress setProgress:0 animated:NO];
         [UIAlertView showWithTitle:NSLocalizedString(@"imageUploadCompleted_title", @"Upload Completed")
-                           message:(self.totalUploadedImages > 1) ? [NSString stringWithFormat:@"%ld %@", self.totalUploadedImages, NSLocalizedString(@"imageImagesUploadCompleted_message", @"images/videos uploaded to your Piwigo server.")]: NSLocalizedString(@"imageImageUploadCompleted_message", @"1 image/video uploaded to your Piwigo server.")
+                           message:(self.totalUploadedImages > 1) ? [NSString stringWithFormat:@"%ld %@", (long)self.totalUploadedImages, NSLocalizedString(@"imageImagesUploadCompleted_message", @"images/videos uploaded to your Piwigo server.")]: NSLocalizedString(@"imageImageUploadCompleted_message", @"1 image/video uploaded to your Piwigo server.")
                  cancelButtonTitle:NSLocalizedString(@"alertOkButton", @"OK")
                  otherButtonTitles:nil
                           tapBlock:nil];
